1_foundations/rag_chat/requirements.txt
openai
openai-agents
pypdf
gradio
python-dotenv
requests
